requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
PyYAML>=6.0.0
gradio>=4.0.0
numpy>=1.24.0
